Make score framerate independent
- Also give player ability to change color using spacebar key

+++ Assets/Resources/Scripts/GameManager.cs
@@ -7,7 +7,7 @@ public class GameManager : MonoBehaviour
     public static GameManager instance;
 
     [HideInInspector]
-    public int HighScore, LastScore;
+    public float HighScore, LastScore;
     public const string HighScorePref = "High Score";
     public float musicVolume, soundVolume;
